move 'mockObject' utility under common, update its unit tests
