+++ src/vs/base/test/common/mockObject.test.ts
@@ -4,88 +4,86 @@
  *--------------------------------------------------------------------------------------------*/
 
 import assert from 'assert';
-import { mockService } from './mock.js';
-import { typeCheck } from '../../../../../base/common/types.js';
-import { randomInt } from '../../../../../base/common/numbers.js';
-import { randomBoolean } from '../../../../../base/test/common/testUtils.js';
-import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../base/test/common/utils.js';
+import { typeCheck } from '../../common/types.js';
+import { randomInt } from '../../common/numbers.js';
+import { mockObject, randomBoolean } from './testUtils.js';
+import { ensureNoDisposablesAreLeakedInTestSuite } from './utils.js';
 
-suite('mockService', () => {
+
+suite('mockObject', () => {
 	ensureNoDisposablesAreLeakedInTestSuite();
 
 	test('• overrides properties and functions', () => {
-		interface ITestService {
-			readonly _serviceBrand: undefined;
-			prop1: string;
-			id: string;
-			readonly counter: number;
-			method1(arg: boolean): string;
-			testMethod2(arg: number): boolean;
+		interface ITestObject {
+			foo: string;
+			bar: string;
+			readonly baz: number;
+			someMethod(arg: boolean): string;
+			anotherMethod(arg: number): boolean;
 		}
 
-		const mock = mockService<ITestService>({
-			id: 'ciao!',
-			counter: 74,
-			testMethod2(arg: number): boolean {
-				return !isNaN(arg);
+		const mock = mockObject<ITestObject>({
+			bar: 'oh hi!',
+			baz: 42,
+			anotherMethod(arg: number): boolean {
+				return isNaN(arg);
 			},
 		});
 
-		typeCheck<ITestService>(mock);
+		typeCheck<ITestObject>(mock);
 
 		assert.strictEqual(
-			mock.id,
-			'ciao!',
-			'id should be overridden',
+			mock.bar,
+			'oh hi!',
+			'bar should be overridden',
 		);
 
 		assert.strictEqual(
-			mock.counter,
-			74,
-			'counter should be overridden',
+			mock.baz,
+			42,
+			'baz should be overridden',
 		);
 
 		assert(
-			mock.testMethod2(randomInt(100)),
+			!(mock.anotherMethod(randomInt(100))),
 			'Must execute overridden method correctly 1.',
 		);
 
 		assert(
-			!(mock.testMethod2(NaN)),
+			mock.anotherMethod(NaN),
 			'Must execute overridden method correctly 2.',
 		);
 
 		assert.throws(() => {
 			// property is not overridden so must throw
 			// eslint-disable-next-line local/code-no-unused-expressions
-			mock.prop1;
+			mock.foo;
 		});
 
 		assert.throws(() => {
 			// function is not overridden so must throw
-			mock.method1(randomBoolean());
+			mock.someMethod(randomBoolean());
 		});
 	});
 
 	test('• immutability of the overrides object', () => {
-		interface ITestService {
-			readonly _serviceBrand: undefined;
+		interface ITestObject {
 			foo: string;
 			bar: string;
-			readonly baz: boolean;
+			readonly baz: number;
 			someMethod(arg: boolean): string;
 			anotherMethod(arg: number): boolean;
 		}
 
-		const overrides: Partial<ITestService> = {
-			baz: false,
+		const overrides: Partial<ITestObject> = {
+			baz: 4,
 		};
-		const mock = mockService<ITestService>(overrides);
-		typeCheck<ITestService>(mock);
+		const mock = mockObject<ITestObject>(overrides);
+		typeCheck<ITestObject>(mock);
 
 		assert.strictEqual(
 			mock.baz,
-			false,
+			4,
 			'baz should be overridden',
 		);
