+++ src/vs/platform/prompts/test/common/utils/mock.ts
@@ -3,43 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { assertOneOf } from '../../../../../base/common/types.js';
-
-/**
- * Mocks an `TObject` with the provided `overrides`.
- *
- * If you need to mock an `Service`, please use {@link mockService}
- * instead which provides better type safety guarantees for the case.
- *
- * @throws Reading non-overridden property or function
- * 		   on `TObject` throws an error.
- */
-export function mockObject<TObject extends Object>(
-	overrides: Partial<TObject>,
-): TObject {
-	// ensure that the overrides object cannot be modified afterward
-	overrides = Object.freeze(overrides);
-
-	const keys = Object.keys(overrides) as (keyof (typeof overrides))[];
-	const service = new Proxy(
-		{},
-		{
-			get: (_target, key: string | number | Symbol) => {
-				// sanity check for the provided `key`
-				assertOneOf(
-					key,
-					keys,
-					`The '${key}' is not mocked.`,
-				);
-
-				return overrides[key];
-			},
-		});
-
-	// note! it's ok to `as TObject` here, because of
-	// 		 the runtime checks in the `Proxy` getter
-	return service as (typeof overrides) as TObject;
-}
+import { mockObject } from '../../../../../base/test/common/testUtils.js';
 
 /**
  * Type for any service.
